# Add controller Service Instance test