--- a/pkg/controller/controller_flow_binding_test.go
+++ b/pkg/controller/controller_flow_binding_test.go
@@ -628,6 +628,3 @@ func TestCreateServiceBindingWithSecretTransform(t *testing.T) {
 		})
 	}
 }
-
-
-