@mszostok 2019-07-02
/test pull-service-catalog-integration