@jberkhahn 2019-07-30
The tests themselves seem good, but I think there's quite a bit of refactoring of the setup stuff that could be done. A lot are dupes of functions we already have in the test/util package.