@jasiu001 2019-08-07
I noticed that all comments are about why I don't use existing functions in `test/util/util.go`. First of all, we moved all controller integration tests from `test/integration` to `pkg/controller` so it is natural we moved also `util` functions.

As you can see in this [commit](https://github.com/kubernetes-sigs/service-catalog/commit/9f31b7d5f8a9b8804b9223272766a6c60d16cf79) which is already merged to master "utils" functions are moved to `controller` package along with controller's tests.
If I am not mistaken, the `test/integration` will be deleted in the future.

In the end, I would like to point out that in "our" `case_test` we use `controllerTest's` methods instead of functions so we don't have to e.g. inject to each "util" function the client, which in my opinion is much easier to read.