@MHBauer 2019-08-13
Is this primarily movement of existing tests or new tests? I do not really have the time to review 1400 lines.

I cannot tell from the commit message or the PR message if this is new or old tests. Moved or recreated or freshly created.