@jasiu001 2019-08-13
@MHBauer this PR is the second part of moved tests (the first part was merged to master, here is the [commit](https://github.com/kubernetes-sigs/service-catalog/commit/9f31b7d5f8a9b8804b9223272766a6c60d16cf79).
As I mentioned above we moved all controller integration tests from `test/integration` to `pkg/controller` and we've rewritten the tests for the new version with CRDs.