@mszostok 2019-08-27
@MHBauer In general here is the issue about that: 
https://github.com/kubernetes-sigs/service-catalog/issues/2597

To make a long story short, current integration tests are tightly coupled with api-server (starting server, etcd). We want to rewrite them into general integration tests with only one dependency to SC controllers.

Right now, the rewritten tests are placed are under `pkg/controller` package with `// +build integration` tag, so they are executed only if you execute 
```
go test --tags="integration" github.com/kubernetes-sigs/service-catalog/pkg/controller/...
```

> @MHBauer: I'll have to dig into the existing tests to find the corresponding other tests.

the test name is exactly the same as defined in https://github.com/kubernetes-sigs/service-catalog/tree/master/test/integration, so it's easy to check them

e.g. `TestServiceBindingDeleteWithAsyncBindInProgress` 

when you execute search:
https://github.com/kubernetes-sigs/service-catalog/search?q=TestServiceBindingDeleteWithAsyncBindInProgress&unscoped_q=TestServiceBindingDeleteWithAsyncBindInProgress

then you will see that test is defined in both places:
- `test/integration/controller_test.go`
- and `pkg/controller/controller_flow_binding_test.go`

they are testing the same logic but the new way does not require the api-server

right now we are in the transition phase and we run both tests (this is also the reason why `pull-service-catalog-integration` takes a little bit more time), so we can be sure that both working well. After merging CRDs we are able to remove the old integration test (with api-server) but still have the same coverage thanks to new tests which were implemented, so we will not regress.