@jasiu001 2019-08-29
@MHBauer considering your comment about parallel tests I paralleled some tests. The list of tests and their execution times before and after are presented in the table below:

test | time execution before | time execution after | notes |
--|--|--|--
TestProvisionInstanceWithRetries | 4.100s |  2.052s |
TestRetryAsyncDeprovision | - | - | single test |
TestServiceInstanceDeleteWithAsyncProvisionInProgress | - | - | already Parallel |
TestServiceInstanceDeleteWithAsyncUpdateInProgress | - | - | already Parallel |
TestCreateServiceInstanceFailsWithNonexistentPlan | - | - | single test |
TestCreateServiceInstanceNonExistentClusterServiceBroker | - | - | single test |
TestCreateServiceInstanceNonExistentClusterServiceClassOrPlan | - | - | - |
TestCreateServiceInstanceWithInvalidParameters | - | - | single test |
TestCreateServiceInstanceWithParameters | 9.288s | 3.104s | |
TestCreateServiceInstanceWithProvisionFailure | 21.232s | 6.048s | |
TestUpdateServiceInstanceChangePlans | 12.521s | 4.266s | |
TestUpdateServiceInstanceChangePlansToNonexistentPlan | - | - | single test |
TestUpdateServiceInstanceNewDashboardResponse | - | - | cannot be run parralel, details in code |
TestUpdateServiceInstanceUpdateParameters | 18.316s | 5.417s | |
**All tests** | 87.120s | 48.008s | |

-----

When it comes to the `TestUpdateServiceInstanceUpdateParameters` test, the problem was related to another test that set the` feature gate` flag to true. The correct performance of the test depended on the order in which the tests were run. The error has been corrected.

-----
I also tried to reproduce failing `TestCreateServiceInstanceWithInvalidParameters` test (with count=100), but each time test passed. I also tried to run this test on @mszostok computer but it passed. Please, I need more details about how do you run this test.

-----
@mszostok missing tests have been added