@jasiu001 2019-08-29
/test pull-service-catalog-test-integration