@MHBauer 2019-08-30
I forgot about feature gates. I remember those preventing nearly any use of parallel. Oh well.