@MHBauer 2019-09-03
Is there a 1second sleep in here somewhere? A lot of tests take that suspiciously specific amount of time.